task compeleted

+++ app/Http/Controllers/UserController.php
@@ -3,11 +3,13 @@ namespace App\Http\Controllers;
 
 use App\Models\order;
 use App\Models\role_user;
+use App\Models\role;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Testing\Fluent\Concerns\Has;
 
 class UserController extends Controller
 {
@@ -43,7 +45,6 @@ class UserController extends Controller
         if ($user) {
             $checkHash = Hash::check($request->password, $user->password);
             if ($checkHash) {
-                // return $user;
                 $user->role;
                 Auth::login($user);
                 return to_route('user.profile', [$user]);
@@ -90,23 +91,35 @@ class UserController extends Controller
 
     public function edit(user $user)
     {
-        return view('admin.user.edit', ['user' => $user]);
+        $roles = role::all();
+        return view('admin.user.edit', ['user' => $user, 'roles'=>$roles]);
     }
 
     public function update(Request $request)
     {
         $user = User::find($request->id);
+        $role = role_user::where('user_id', $user->id)->delete();
+        role_user::create([
+            'user_id'=>$user->id,
+            'role_id'=>$request->role
+        ]);
         $user->name = $request->name;
         $user->family = $request->family;
-        $user->phoneNumber = $request->phoneNumber;
-        $user->email = $request->email;
+        if (isset($request->phoneNumber)) {
+            $user->phoneNumber = $request->phoneNumber;
+        }
+        if (isset($request->email)) {
+            $user->email = $request->email;
+        }
 
         if ($request->password) {
             $password = Hash::make($request->password);
             $user->password = $password;
         }
         if ($request->main_image) {
-            Storage::disk('public')->delete($user->main_image);
+            if ($user->main_image) {
+                Storage::disk('public')->delete($user->main_image);
+            }
             $name = $request->main_image->getClientOriginalName();
             $fullName = time() . '_' . $name;
             $path = $request->file('main_image')->storeAs('images', $fullName, 'public');
@@ -170,5 +183,33 @@ class UserController extends Controller
     public function checkAuth(Request $request){
         $user = User::where('phoneNumber', $request->phoneNumber)->first();
         return response()->json($user);
+    }
+
+    public function create_user(){
+        $roles = role::all();
+        return view('admin.user.create', ['roles'=>$roles]);
+    }
+
+    public function store_user(Request $request){
+        $password = Hash::make($request->password);
+        $path = null;
+        if (isset($request->main_image)) {
+            $name = $request->main_image->getClientOriginalName();
+            $fullName = time()."_".$name;
+            $path = $request->file('main_image')->storeAs('images', $fullName, 'public');
+        }
+        $user_id = User::insertGetId([
+            'name'=>$request->name,
+            'family'=>$request->family,
+            'phoneNumber'=>$request->phoneNumber,
+            'email'=>$request->email,
+            'main_image'=>$path,
+            'password'=>$password,
+        ]);
+        role_user::create([
+            'user_id'=>$user_id,
+            'role_id'=>$request->role
+        ]);
+        return to_route('user.list');
     }
 }
